feat: refactor component imports and create new SignInForm, SignUpForm, DashboardPageView, HeroSection, and LandingPageView components
