+++ app/dashboard/page.tsx
@@ -1,4 +1,4 @@
-import DashboardPageView from "@/components/dashboard/dashboard-page-view";
+import DashboardPageView from "@/modules/dashboard/ui/views/dashboard-page-view";
 import { auth } from "@/lib/auth";
 import { headers } from "next/headers";
 import { redirect } from "next/navigation";
