+++ modules/landing/ui/views/landing-page-view.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+import HeroSection from "../components/hero-section";
+
+const LandingPageView = () => {
+  return (
+    <section>
+      <HeroSection />
+    </section>
+  );
+};
+
+export default LandingPageView;
